# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Scientific Workbench API",
    "version": "0.2.0"
  },
  "paths": {
    "/api/v1/documents/{id}/blocks/{blockId}/resolve-data": {
      "post": {
        "operationId": "document_resolve_data_identity",
        "description": "数据区块身份丢失后明确关联已有Data（加载当前内容）或新建，保留其他正文",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "blockId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dataId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "createNew": {
                    "type": "boolean"
                  },
                  "expectedVersion": {
                    "type": "integer",
                    "minimum": 1
                  }
                },
                "required": [
                  "expectedVersion"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作结果，实体包含实际版本；分页请求返回items/total/offset/limit"
          },
          "401": {
            "description": "凭据无效"
          },
          "403": {
            "description": "缺少权限"
          },
          "409": {
            "description": "版本或幂等冲突"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/workspace/status": {
      "get": {
        "operationId": "workspace_status",
        "description": "读取目录、实体数量与待更新状态",
        "parameters": [],
        "responses": {
          "200": {
            "description": "操作结果，实体包含实际版本；分页请求返回items/total/offset/limit"
          },
          "401": {
            "description": "凭据无效"
          },
          "403": {
            "description": "缺少权限"
          },
          "409": {
            "description": "版本或幂等冲突"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/workspace/settings": {
      "get": {
        "operationId": "workspace_settings",
        "description": "读取工作区公开设置，不返回凭据",
        "parameters": [],
        "responses": {
          "200": {
            "description": "操作结果，实体包含实际版本；分页请求返回items/total/offset/limit"
          },
          "401": {
            "description": "凭据无效"
          },
          "403": {
            "description": "缺少权限"
          },
          "409": {
            "description": "版本或幂等冲突"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/workspace/rebuild": {
      "post": {
        "operationId": "workspace_rebuild",
        "description": "从业务文件重建索引",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作结果，实体包含实际版本；分页请求返回items/total/offset/limit"
          },
          "401": {
            "description": "凭据无效"
          },
          "403": {
            "description": "缺少权限"
          },
          "409": {
            "description": "版本或幂等冲突"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/samples": {
      "get": {
        "operationId": "sample_list",
        "description": "分页列出样品属性和提取状态",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作结果，实体包含实际版本；分页请求返回items/total/offset/limit"
          },
          "401": {
            "description": "凭据无效"
          },
          "403": {
            "description": "缺少权限"
          },
          "409": {
            "description": "版本或幂等冲突"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "operationId": "sample_create",
        "description": "新建独立样品正文",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "body": {
                    "type": "string"
                  }
                },
                "required": []
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作结果，实体包含实际版本；分页请求返回items/total/offset/limit"
          },
          "401": {
            "description": "凭据无效"
          },
          "403": {
            "description": "缺少权限"
          },
          "409": {
            "description": "版本或幂等冲突"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/samples/{id}": {
      "get": {
        "operationId": "sample_get",
        "description": "读取样品正文、属性和实际内容版本；latest=true先完成提取",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "latest",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作结果，实体包含实际版本；分页请求返回items/total/offset/limit"
          },
          "401": {
            "description": "凭据无效"
          },
          "403": {
            "description": "缺少权限"
          },
          "409": {
            "description": "版本或幂等冲突"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "operationId": "sample_update",
        "description": "按版本修改样品编号或标题",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "expectedVersion": {
                    "type": "integer",
                    "minimum": 1
                  }
                },
                "required": [
                  "expectedVersion"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作结果，实体包含实际版本；分页请求返回items/total/offset/limit"
          },
          "401": {
            "description": "凭据无效"
          },
          "403": {
            "description": "缺少权限"
          },
          "409": {
            "description": "版本或幂等冲突"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/samples/{id}/copy": {
      "post": {
        "operationId": "sample_copy",
        "description": "复制操作模板，排除Data、附件和论点结果",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string"
                  }
                },
                "required": []
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作结果，实体包含实际版本；分页请求返回items/total/offset/limit"
          },
          "401": {
            "description": "凭据无效"
          },
          "403": {
            "description": "缺少权限"
          },
          "409": {
            "description": "版本或幂等冲突"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/samples/batch": {
      "post": {
        "operationId": "sample_batch",
        "description": "按模板逐行修改属性并批量创建独立样品",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "templateId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "expectedVersion": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "rows": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "values": {
                          "type": "object"
                        }
                      },
                      "required": [
                        "values"
                      ]
                    }
                  }
                },
                "required": [
                  "templateId",
                  "expectedVersion",
                  "rows"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作结果，实体包含实际版本；分页请求返回items/total/offset/limit"
          },
          "401": {
            "description": "凭据无效"
          },
          "403": {
            "description": "缺少权限"
          },
          "409": {
            "description": "版本或幂等冲突"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/documents/{id}/finalize": {
      "post": {
        "operationId": "document_finalize",
        "description": "整篇提取已保存正文，重复调用更新原记录",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作结果，实体包含实际版本；分页请求返回items/total/offset/limit"
          },
          "401": {
            "description": "凭据无效"
          },
          "403": {
            "description": "缺少权限"
          },
          "409": {
            "description": "版本或幂等冲突"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/attachments/{id}": {
      "get": {
        "operationId": "attachment_get",
        "description": "读取附件元数据和稳定位置",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "操作结果，实体包含实际版本；分页请求返回items/total/offset/limit"
          },
          "401": {
            "description": "凭据无效"
          },
          "403": {
            "description": "缺少权限"
          },
          "409": {
            "description": "版本或幂等冲突"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}